Optimized CustomItemRelationTestCase
